Add files via upload
The function takes two matrices as input. One with independent variables in each column and the second one with only one column containing the dependent variable. It conducts leave-one-out cross-validation and gives mean values of coefficients, mean error of estimation, mean adjusted r-squared, and mean p-value of the whole model.